Side tracked - example cards
Found a good way to integrate the existing <ColorCard /> inside each of the navigation links on the "/convert" page. Each of these links is much more complex, but they hold somewhat randomized examples of the values to expect when you convert a certain color.

This side track has given me a jump start on the form pages though as a lot of the logic has already been figured out.

Will still need to create the separate <ColorBadge />s.

I should find a way to randomize the example color each time it is toggled on.