finished
Finished the initial goal of being able to convert many formats of colors. The code at this point isn't as clever as I dreamed. Maybe I'll attempt to DRY it out later, or maybe it won't be worth the complexity to remove all of the "repeated" files since certain colors need special attention.

This site could be extended to include theming solutions or other things.